easy to get types
