+++ lib/Joypad.d.ts
@@ -1,5 +1,4 @@
-import JoypadEventEmitter from './JoypadEventEmitter';
-import { ButtonState, StickState } from './generate-button-state';
+import { JoypadEventEmitter } from './JoypadEventEmitter';
 import type { JoypadConfig } from './JoypadManager';
 import type { JoypadMap } from './mappings';
 export interface VibrationParameters {
@@ -8,14 +7,26 @@ export interface VibrationParameters {
     weakMagnitude: number;
     strongMagnitude: number;
 }
-export interface VibrationActuator {
+interface VibrationActuator {
     playEffect: (type: 'dual-rumble', parameters: VibrationParameters) => Promise<'invalid-parameter' | 'complete' | 'preempted'>;
     reset: () => Promise<'complete'>;
 }
 interface Gamepad extends globalThis.Gamepad {
     vibrationActuator?: VibrationActuator;
 }
-export default class Joypad extends JoypadEventEmitter {
+export interface StickState {
+    name: string;
+    value: {
+        x: number;
+        y: number;
+        angle: number;
+    };
+}
+export interface ButtonState {
+    name: string;
+    value: number;
+}
+export declare class Joypad extends JoypadEventEmitter {
     readonly index: number;
     private joypadConfig;
     private mappings;
