+++ lib/Joypad.js
@@ -7,10 +7,10 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
-import JoypadEventEmitter from './JoypadEventEmitter';
+import { JoypadEventEmitter } from './JoypadEventEmitter';
 import * as JOYPAD_EVENTS from './event-names';
 import generateButtonState from './generate-button-state';
-export default class Joypad extends JoypadEventEmitter {
+export class Joypad extends JoypadEventEmitter {
     /**
      * @param index the gamepad index in the JoypadManager.joypads array
      * @param joypadConfig the joypad config
