+++ lib/JoypadEventEmitter.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-duplicates */
 import JoypadEventTracker from './JoypadEventTracker';
 import * as JOYPAD_EVENTS from './event-names';
 function generateEvents() {
@@ -10,7 +11,7 @@ function generateEvents() {
         [JOYPAD_EVENTS.STICK_MOVE]: new JoypadEventTracker(JOYPAD_EVENTS.STICK_MOVE),
     };
 }
-export default class JoypadEventEmitter {
+export class JoypadEventEmitter {
     constructor() {
         this.events = generateEvents();
     }
