+++ lib/JoypadManager.js
@@ -1,5 +1,5 @@
-import Joypad from './Joypad';
-export default class JoypadManager {
+import { Joypad } from './Joypad';
+export class JoypadManager {
     /**
      * Initiate the JoypadManager
      * @param joypadConfig Optional Joypad configuration
