+++ lib/generate-button-state.d.ts
@@ -1,16 +1,5 @@
 import { JoypadMap } from './mappings';
-export interface StickState {
-    name: string;
-    value: {
-        x: number;
-        y: number;
-        angle: number;
-    };
-}
-export interface ButtonState {
-    name: string;
-    value: number;
-}
+import type { ButtonState, StickState } from './Joypad';
 export default function generateButtonState(id: string, customMappings: JoypadMap[]): {
     mapping: JoypadMap | undefined;
     buttons: ButtonState[];
