+++ lib/index.d.ts
@@ -1,2 +1,4 @@
-import JoypadManager from './JoypadManager';
+import { JoypadManager } from './JoypadManager';
+import * as Types from './types';
 export default JoypadManager;
+export { Types };
