+++ lib/index.js
@@ -1,3 +1,5 @@
-import JoypadManager from './JoypadManager';
+import { JoypadManager } from './JoypadManager';
+import * as Types from './types';
 export default JoypadManager;
+export { Types };
 //# sourceMappingURL=index.js.map
