+++ lib/types.d.ts
@@ -1,0 +1,5 @@
+import * as Joypad from './Joypad';
+import * as JoypadManager from './JoypadManager';
+import * as Events from './JoypadEventEmitter';
+import * as Mapping from './mappings';
+export { Joypad, JoypadManager, Events, Mapping, };
