+++ lib/types.js
@@ -1,0 +1,6 @@
+import * as Joypad from './Joypad';
+import * as JoypadManager from './JoypadManager';
+import * as Events from './JoypadEventEmitter';
+import * as Mapping from './mappings';
+export { Joypad, JoypadManager, Events, Mapping, };
+//# sourceMappingURL=types.js.map
