+++ src/index.ts
@@ -1,3 +1,6 @@
-import JoypadManager from './JoypadManager';
+import { JoypadManager } from './JoypadManager';
+import * as Types from './types';
 
 export default JoypadManager;
+
+export { Types };
