+++ tsconfig.json
@@ -11,7 +11,7 @@
     "moduleResolution": "Node",
     "resolveJsonModule": true,
     "declaration": true,
-      "lib": [
+    "lib": [
       "DOM",
       "DOM.Iterable",
       "ESNext"
